refactor(server): extract tool arguments, result helpers and tool schemas into src/server/
First code move of the src/server.ts split, behind the tools/list snapshot
fence. Three pure modules, bodies moved verbatim:

- src/server/tool-arguments.ts: ToolArguments, TOOL_ARG_EXPECTED_TYPES,
  coerceToolArguments. server.ts re-exports coerceToolArguments so the
  existing `import { coerceToolArguments } from 'src/server.js'` keeps working.
- src/server/tool-result.ts: the ToolResult shape every handler produces,
  jsonResult / prettyJsonResult (for the handler modules that follow), and
  sanitizeRequest / extractPayloadSuccess lifted off the class unchanged.
- src/server/tool-schemas.ts: buildToolDefinitions({ supportedLanguages,
  environment }) and getPathDescription(environment, name). The ListTools
  handler on the class now logs, discovers languages, and returns
  { tools: buildToolDefinitions(...) } with this.environment read at request
  time, exactly as before.

src/server.ts header notes that src/server/ is a sibling directory (NodeNext
resolves ./server.js and ./server/x.js without conflict).
